requirements.txt
beautifulsoup4
click
faiss-cpu
numpy
packaging
python-dotenv
requests
langchain-text-splitters